figma-transformer: split Kiwi parser contracts
