+++ figma-transformer/tests/contract/run.php
@@ -10,6 +10,7 @@ require_once __DIR__ . '/EffectsContract.php';
 require_once __DIR__ . '/FixtureMatrixContract.php';
 require_once __DIR__ . '/GeometryBoxContract.php';
 require_once __DIR__ . '/KiwiSkippedFieldInventoryContract.php';
+require_once __DIR__ . '/KiwiParserContract.php';
 require_once __DIR__ . '/LayoutMismatchContract.php';
 require_once __DIR__ . '/NodeTraceContract.php';
 require_once __DIR__ . '/OriginInferenceContract.php';
@@ -22,11 +23,7 @@ require_once __DIR__ . '/VectorCommandBlobContract.php';
 require_once __DIR__ . '/VectorRenderingContract.php';
 require_once __DIR__ . '/VisualNodeMapContract.php';
 
-use Automattic\BlocksEngine\FigmaTransformer\Compression\ZstdCapability;
-use Automattic\BlocksEngine\FigmaTransformer\Compression\ZstdCommandDecoder;
-use Automattic\BlocksEngine\FigmaTransformer\FigFile\FigArchiveReader;
 use Automattic\BlocksEngine\FigmaTransformer\FigFile\FigKiwiDecoder;
-use Automattic\BlocksEngine\FigmaTransformer\FigFile\FigKiwiParser;
 use Automattic\BlocksEngine\FigmaTransformer\Parity\ParityReportBuilder;
 use Automattic\BlocksEngine\FigmaTransformer\Parity\VisualAttributionReportBuilder;
 use Automattic\BlocksEngine\FigmaTransformer\Scenegraph\ScenegraphFrameClassifier;
@@ -1309,395 +1306,7 @@ if ( function_exists('imagecreatetruecolor') && function_exists('imagepng') ) {
     $assert(array('background', 'positioned') === ($visualAttribution['top_nodes'][0]['features'] ?? null), 'visual-attribution-node-features');
 }
 
-$fixture = blocks_engine_figma_transformer_create_fig_wrapper_fixture();
-$fileResult = blocks_engine_figma_transformer_transform_file($fixture);
-@unlink($fixture);
-
-$canvas = $fileResult['source_reports']['figma']['archive']['canvas'] ?? array();
-$chunks = $canvas['chunks'] ?? array();
-$diagnosticCodes = array_map(
-    static fn (array $diagnostic): string => (string) ($diagnostic['code'] ?? ''),
-    $fileResult['diagnostics'] ?? array()
-);
-$zstdCapability = new ZstdCapability();
-$zstdStatus = $zstdCapability->status();
-$zstdCapabilityDiagnostic = $zstdCapability->diagnostic('ContractTest', 0);
-$zstdCapabilityCode = (string) ($zstdCapabilityDiagnostic['code'] ?? '');
-$zstdDiagnostic = null;
-foreach ( $fileResult['diagnostics'] ?? array() as $diagnostic ) {
-    if ( $zstdCapabilityCode === ($diagnostic['code'] ?? null) ) {
-        $zstdDiagnostic = $diagnostic;
-        break;
-    }
-}
-
-$assert('success_with_warnings' === ($fileResult['status'] ?? null), 'file-transform-status');
-$assert('fig-kiwi' === ($canvas['prelude'] ?? null), 'fig-kiwi-prelude');
-$assert(106 === ($canvas['version'] ?? null), 'fig-kiwi-version');
-$assert('inner.fig' === ($fileResult['source_reports']['figma']['input']['nested_fig'] ?? null), 'wrapper-nested-fig');
-$assert(5 === count($chunks), 'fig-kiwi-chunk-count');
-$assert('zlib' === ($chunks[0]['compression'] ?? null), 'fig-kiwi-first-chunk-zlib');
-$assert('json' === ($chunks[0]['payload']['classification'] ?? null), 'fig-kiwi-first-chunk-json');
-$assert(isset($chunks[0]['payload']['json']['nodes']), 'fig-kiwi-first-chunk-nodes-candidate');
-$assert('json_invalid' === ($chunks[1]['payload']['classification'] ?? null), 'fig-kiwi-second-chunk-json-invalid');
-$assert(isset($chunks[2]['payload']['json']['NODE_CHANGES']), 'fig-kiwi-third-chunk-node-changes');
-$assert('binary' === ($chunks[3]['payload']['classification'] ?? null), 'fig-kiwi-fourth-chunk-binary');
-$assert('zstd' === ($chunks[4]['compression'] ?? null), 'fig-kiwi-fifth-chunk-zstd');
-$assert(in_array($zstdCapabilityCode, $diagnosticCodes, true), 'fig-kiwi-zstd-capability-diagnostic');
-$assert(is_bool($zstdStatus['available'] ?? null), 'zstd-status-available-bool');
-$assert(is_bool($zstdStatus['extension_loaded'] ?? null), 'zstd-status-extension-loaded-bool');
-$assert(is_array($zstdStatus['functions'] ?? null), 'zstd-status-functions-array');
-$assert(array_key_exists('zstd_uncompress', $zstdStatus['functions'] ?? array()), 'zstd-status-uncompress-function');
-$assert(array_key_exists('adapter_registered', $zstdStatus), 'zstd-status-adapter-registered-key');
-$assert(array_key_exists('wordpress_filter_registered', $zstdStatus), 'zstd-status-wordpress-filter-registered-key');
-$assert(($zstdStatus['available'] ?? null) === (null !== ($zstdStatus['provider'] ?? null)), 'zstd-status-available-matches-provider');
-$assert(($zstdStatus['available'] ?? null) === ($zstdDiagnostic['context']['available'] ?? null), 'fig-kiwi-zstd-diagnostic-availability-context');
-if ( true === ($zstdStatus['available'] ?? false) && function_exists('zstd_compress') ) {
-    $zstdCompressed = zstd_compress('contract zstd round trip');
-    $zstdRoundTrip = false !== $zstdCompressed ? $zstdCapability->uncompress($zstdCompressed, 'ContractTest', 1) : array('data' => null, 'diagnostics' => array());
-    $assert('contract zstd round trip' === ($zstdRoundTrip['data'] ?? null), 'zstd-real-round-trip');
-    $assert(isset($chunks[4]['inflated_bytes']), 'fig-kiwi-zstd-real-fixture-inflated');
-} else {
-    $zstdUnavailable = $zstdCapability->uncompress("\x28\xb5\x2f\xfd" . 'synthetic-zstd-frame', 'ContractTest', 1);
-    $assert(null === ($zstdUnavailable['data'] ?? null), 'zstd-unavailable-returns-null');
-    $assert(in_array((string) ($zstdUnavailable['diagnostics'][0]['code'] ?? ''), array('figma_transformer_zstd_extension_missing', 'figma_transformer_zstd_function_missing'), true), 'zstd-unavailable-diagnostic-code');
-}
-
-$adapterCapability = new ZstdCapability(static function (string $payload): string|false {
-    if ( "\x28\xb5\x2f\xfd" !== substr($payload, 0, 4) ) {
-        return false;
-    }
-
-    return json_encode(array('NODE_CHANGES' => array()), JSON_THROW_ON_ERROR);
-});
-$adapterStatus = $adapterCapability->status();
-$adapterResult = $adapterCapability->uncompress("\x28\xb5\x2f\xfd" . 'adapter-frame', 'ContractTest', 2);
-$adapterCanvasResult = ( new FigKiwiParser($adapterCapability) )->parse(
-    'fig-kiwi'
-    . pack('V', 106)
-    . blocks_engine_figma_transformer_kiwi_chunk("\x28\xb5\x2f\xfd" . 'adapter-frame')
-);
-$failingAdapterResult = ( new ZstdCapability(static fn (): false => false) )->uncompress("\x28\xb5\x2f\xfd" . 'adapter-frame', 'ContractTest', 3);
-$commandAdapterResult = ( new ZstdCapability(new ZstdCommandDecoder(array(PHP_BINARY, '-r', '$payload = stream_get_contents(STDIN); fwrite(STDOUT, $payload);'))) )->uncompress('command adapter bytes', 'ContractTest', 4);
-
-$assert(true === ($adapterStatus['available'] ?? null), 'zstd-adapter-status-available');
-$assert('adapter' === ($adapterStatus['provider'] ?? null) || 'ext-zstd' === ($adapterStatus['provider'] ?? null), 'zstd-adapter-status-provider');
-$assert('{"NODE_CHANGES":[]}' === ($adapterResult['data'] ?? null), 'zstd-adapter-decodes-payload');
-$assert('json' === ($adapterCanvasResult['canvas']['chunks'][0]['payload']['classification'] ?? null), 'fig-kiwi-zstd-adapter-classifies-json');
-$assert('figma_transformer_zstd_adapter_failed' === ($failingAdapterResult['diagnostics'][0]['code'] ?? null), 'zstd-adapter-failure-diagnostic');
-$assert('command adapter bytes' === ($commandAdapterResult['data'] ?? null), 'zstd-command-adapter-decodes-payload');
-$assert('figma_transformer_zstd_command_used' === ($commandAdapterResult['diagnostics'][1]['code'] ?? null), 'zstd-command-adapter-diagnostic');
-$assert(! empty($fileResult['files']), 'file-transform-renders-decoded-scenegraph');
-$assert(4 === ($fileResult['metrics']['node_count'] ?? null), 'file-transform-node-count');
-$assert(2 === ($fileResult['metrics']['decoded_payload_candidate_count'] ?? null), 'file-transform-decoded-candidate-count');
-$assert(2 === ($fileResult['metrics']['selected_decoded_payload_index'] ?? null), 'file-transform-selected-node-changes-index');
-$assert('NODE_CHANGES' === ($fileResult['source_reports']['figma']['decoded_scenegraph']['shape'] ?? null), 'file-transform-selected-node-changes-shape');
-$assert(isset($fileResult['source_reports']['figma']['html']), 'file-transform-html-source-report');
-$assert('blocks-engine/figma-transformer/compiled-site/v1' === ($fileResult['source_reports']['compiled_site']['schema'] ?? null), 'file-transform-compiled-site-source-report');
-$assert('synthetic' === ($fileResult['source_reports']['figma']['assets'][0]['id'] ?? null), 'archive-asset-id');
-$assert('images/synthetic' === ($fileResult['source_reports']['figma']['assets'][0]['path'] ?? null), 'archive-asset-path');
-$assert('asset' === ($fileResult['source_reports']['figma']['assets'][0]['content'] ?? null), 'archive-asset-content');
-$assert('assets/synthetic.bin' === ($fileResult['assets'][0]['path'] ?? null), 'archive-asset-emitted-from-decoded-scenegraph');
-
-$assetMetadataFixture = SyntheticFigKiwiFixtureBuilder::figArchive(
-    SyntheticFigKiwiFixtureBuilder::canvas(array(SyntheticFigKiwiFixtureBuilder::jsonZlibChunk(array('metadata' => array('ignored' => true))))),
-    array('images/metadata-only' => "\x89PNG\r\n\x1a\n" . str_repeat('asset-bytes', 20))
-);
-$assetMetadataResult = ( new FigArchiveReader() )->read($assetMetadataFixture, array('include_asset_content' => false));
-@unlink($assetMetadataFixture);
-$assert('metadata-only' === ($assetMetadataResult['assets'][0]['id'] ?? null), 'archive-asset-metadata-id');
-$assert('image/png' === ($assetMetadataResult['assets'][0]['mime_type'] ?? null), 'archive-asset-metadata-sniffs-mime');
-$assert(! array_key_exists('content', $assetMetadataResult['assets'][0] ?? array()), 'archive-asset-metadata-omits-content');
-
-$pendingFixture = blocks_engine_figma_transformer_create_pending_decoder_fig_wrapper_fixture();
-$pendingResult = blocks_engine_figma_transformer_transform_file($pendingFixture);
-@unlink($pendingFixture);
-$pendingDiagnosticCodes = array_map(
-    static fn (array $diagnostic): string => (string) ($diagnostic['code'] ?? ''),
-    $pendingResult['diagnostics'] ?? array()
-);
-$assert('unsupported_decoder_pending' === ($pendingResult['status'] ?? null), 'pending-decoder-status');
-$assert(in_array('figma_transformer_decoded_scenegraph_missing', $pendingDiagnosticCodes, true), 'pending-decoder-diagnostic');
-
-$kiwiSchemaBytes = blocks_engine_figma_transformer_kiwi_schema_fixture();
-$kiwiMessageBytes = blocks_engine_figma_transformer_kiwi_message_fixture();
-$kiwiDecoder = new FigKiwiDecoder();
-$kiwiSchemaResult = $kiwiDecoder->decodeSchema($kiwiSchemaBytes);
-$kiwiMessageResult = $kiwiDecoder->decodeMessage($kiwiMessageBytes, $kiwiSchemaResult['schema'] ?? array());
-$assert(null !== ($kiwiSchemaResult['schema'] ?? null), 'kiwi-schema-decodes');
-$assert('NODE_CHANGES' === ($kiwiMessageResult['message']['type'] ?? null), 'kiwi-message-enum-decodes');
-$assert(array('alpha', 'beta') === ($kiwiMessageResult['message']['nodeChanges'] ?? null), 'kiwi-message-array-decodes');
-
-$injectedParser = new FigKiwiParser(new ZstdCapability(static fn (string $payload, array $context): string => $kiwiMessageBytes));
-$injectedCanvas = $injectedParser->parse(
-    'fig-kiwi'
-    . pack('V', 106)
-    . blocks_engine_figma_transformer_kiwi_chunk(gzdeflate($kiwiSchemaBytes))
-    . blocks_engine_figma_transformer_kiwi_chunk("\x28\xb5\x2f\xfd" . 'synthetic-zstd-frame')
-);
-$injectedChunks = $injectedCanvas['canvas']['chunks'] ?? array();
-$injectedDiagnosticCodes = array_map(
-    static fn (array $diagnostic): string => (string) ($diagnostic['code'] ?? ''),
-    $injectedCanvas['diagnostics'] ?? array()
-);
-$assert('kiwi_schema' === ($injectedChunks[0]['payload']['classification'] ?? null), 'kiwi-parser-classifies-schema');
-$assert('kiwi_message' === ($injectedChunks[1]['payload']['classification'] ?? null), 'kiwi-parser-classifies-message');
-$assert('NODE_CHANGES' === ($injectedChunks[1]['payload']['kiwi_message']['type'] ?? null), 'kiwi-parser-message-type');
-$assert(in_array('figma_transformer_zstd_adapter_available', $injectedDiagnosticCodes, true), 'zstd-injected-decoder-diagnostic');
-
-$guardedCanvas = ( new FigKiwiParser(new ZstdCapability(static fn (string $payload, array $context): string => $kiwiMessageBytes)) )->parse(
-    'fig-kiwi'
-    . pack('V', 106)
-    . blocks_engine_figma_transformer_kiwi_chunk(gzdeflate($kiwiSchemaBytes))
-    . blocks_engine_figma_transformer_kiwi_chunk("\x28\xb5\x2f\xfd" . 'synthetic-zstd-frame'),
-    array('max_kiwi_message_decode_bytes' => 1)
-);
-$guardedChunks = $guardedCanvas['canvas']['chunks'] ?? array();
-$guardedDiagnosticCodes = array_map(
-    static fn (array $diagnostic): string => (string) ($diagnostic['code'] ?? ''),
-    $guardedCanvas['diagnostics'] ?? array()
-);
-$assert('kiwi_message' === ($guardedChunks[1]['payload']['classification'] ?? null), 'kiwi-parser-selectively-decodes-oversized-message');
-$assert('selective' === ($guardedChunks[1]['payload']['kiwi_message_decode'] ?? null), 'kiwi-parser-selective-message-mode');
-$assert(in_array('figma_transformer_kiwi_message_selective_decode_used', $guardedDiagnosticCodes, true), 'kiwi-parser-selective-message-diagnostic');
-
-$kiwiFrameMaskDecoder = new FigKiwiDecoder();
-$kiwiFrameMaskSchema = $kiwiFrameMaskDecoder->decodeSchema(blocks_engine_figma_transformer_kiwi_frame_mask_schema_fixture());
-$kiwiFrameMaskMessage = $kiwiFrameMaskDecoder->decodeMessageSelective(
-    blocks_engine_figma_transformer_kiwi_frame_mask_message_fixture(),
-    $kiwiFrameMaskSchema['schema'] ?? array()
-);
-$kiwiFrameMaskNode = $kiwiFrameMaskMessage['message']['nodeChanges'][0] ?? array();
-$assert(false === ($kiwiFrameMaskNode['frameMaskDisabled'] ?? null), 'kiwi-selective-decodes-frame-mask-disabled');
-
-$kiwiDerivedTextSchema = $kiwiDecoder->decodeSchema(blocks_engine_figma_transformer_kiwi_derived_text_schema_fixture());
-$kiwiDerivedTextMessage = $kiwiDecoder->decodeMessageSelective(
-    blocks_engine_figma_transformer_kiwi_derived_text_message_fixture(),
-    $kiwiDerivedTextSchema['schema'] ?? array()
-);
-$kiwiDerivedText = $kiwiDerivedTextMessage['message']['nodeChanges'][0]['derivedTextData'] ?? array();
-$assert(120.0 === ($kiwiDerivedText['layoutSize']['x'] ?? null), 'kiwi-selective-decodes-derived-text-layout-width');
-$assert(1 === count($kiwiDerivedText['baselines'] ?? array()), 'kiwi-selective-decodes-derived-text-baselines');
-$assert(24.0 === ($kiwiDerivedText['baselines'][0]['lineHeight'] ?? null), 'kiwi-selective-decodes-derived-text-baseline-line-height');
-$assert(! array_key_exists('glyphs', $kiwiDerivedText), 'kiwi-selective-skips-derived-text-glyphs');
-$assert('Inter' === ($kiwiDerivedText['fontMetaData']['key']['family'] ?? null), 'kiwi-selective-decodes-derived-text-font-family');
-$assert(700 === ($kiwiDerivedText['fontMetaData']['fontWeight'] ?? null), 'kiwi-selective-decodes-derived-text-font-weight');
-
-$kiwiFrameMaskResult = blocks_engine_figma_transformer_transform_scenegraph(array(
-    'document' => array(
-        'id'       => 'kiwi-mask:document',
-        'type'     => 'DOCUMENT',
-        'name'     => 'Document',
-        'children' => array(
-            array(
-                'id'                => 'kiwi-mask:frame',
-                'type'              => 'FRAME',
-                'name'              => 'Kiwi Mask Frame',
-                'width'             => 120,
-                'height'            => 80,
-                'frameMaskDisabled' => false,
-            ),
-        ),
-    ),
-));
-$kiwiFrameMaskDisabledResult = blocks_engine_figma_transformer_transform_scenegraph(array(
-    'document' => array(
-        'id'       => 'kiwi-mask:disabled-document',
-        'type'     => 'DOCUMENT',
-        'name'     => 'Document',
-        'children' => array(
-            array(
-                'id'                => 'kiwi-mask:disabled-frame',
-                'type'              => 'FRAME',
-                'name'              => 'Kiwi Disabled Mask Frame',
-                'width'             => 120,
-                'height'            => 80,
-                'frameMaskDisabled' => true,
-            ),
-        ),
-    ),
-));
-$kiwiFrameMaskCss = $fileContent($kiwiFrameMaskResult, 'style.css');
-$kiwiFrameMaskDisabledCss = $fileContent($kiwiFrameMaskDisabledResult, 'style.css');
-$assert(str_contains($kiwiFrameMaskCss, '.figma-node-kiwi-mask-frame-kiwi-mask-frame{width:120px;height:80px;overflow:hidden}'), 'kiwi-frame-mask-disabled-false-clips-content');
-$assert(str_contains($kiwiFrameMaskDisabledCss, '.figma-node-kiwi-mask-disabled-frame-kiwi-disabled-mask-frame{width:120px;height:80px}'), 'kiwi-frame-mask-disabled-true-does-not-clip');
-
-$wirePayload = SyntheticFigKiwiFixtureBuilder::sampleWirePayload();
-$wireCanvasResult = ( new FigKiwiParser() )->parse(
-    SyntheticFigKiwiFixtureBuilder::canvas(array(SyntheticFigKiwiFixtureBuilder::zlibChunk($wirePayload)))
-);
-$wire = $wireCanvasResult['canvas']['chunks'][0]['payload']['wire'] ?? array();
-$wireRecords = $wire['records'] ?? array();
-
-$assert('binary' === ($wireCanvasResult['canvas']['chunks'][0]['payload']['classification'] ?? null), 'fig-kiwi-wire-payload-remains-binary');
-$assert('protobuf_wire' === ($wire['format'] ?? null), 'fig-kiwi-wire-format');
-$assert(true === ($wire['complete'] ?? null), 'fig-kiwi-wire-complete');
-$assert(3 === ($wire['record_count'] ?? null), 'fig-kiwi-wire-record-count');
-$assert(1 === ($wireRecords[0]['field_number'] ?? null), 'fig-kiwi-wire-varint-field-number');
-$assert(0 === ($wireRecords[0]['wire_type'] ?? null), 'fig-kiwi-wire-varint-type');
-$assert(150 === ($wireRecords[0]['value'] ?? null), 'fig-kiwi-wire-varint-value');
-$assert('hello' === ($wireRecords[1]['text_preview'] ?? null), 'fig-kiwi-wire-length-text-preview');
-$assert('01020304' === ($wireRecords[2]['preview_hex'] ?? null), 'fig-kiwi-wire-fixed32-preview');
-
-$unknownBinaryResult = ( new FigKiwiParser() )->parse(
-    SyntheticFigKiwiFixtureBuilder::canvas(array(SyntheticFigKiwiFixtureBuilder::zlibChunk("\x00\x01\x02unknown")))
-);
-$unknownBinaryPayload = $unknownBinaryResult['canvas']['chunks'][0]['payload'] ?? array();
-$assert('binary' === ($unknownBinaryPayload['classification'] ?? null), 'fig-kiwi-unknown-binary-classification');
-$assert(10 === ($unknownBinaryPayload['bytes'] ?? null), 'fig-kiwi-unknown-binary-byte-count');
-$assert('000102756e6b6e6f776e' === ($unknownBinaryPayload['preview_hex'] ?? null), 'fig-kiwi-unknown-binary-preview');
-$assert('zero_field_key' === ($unknownBinaryPayload['wire']['reason'] ?? null), 'fig-kiwi-unknown-binary-wire-stop-reason');
-
-$truncatedVarintResult = ( new FigKiwiParser() )->parse(
-    SyntheticFigKiwiFixtureBuilder::canvas(array(SyntheticFigKiwiFixtureBuilder::zlibChunk(SyntheticFigKiwiFixtureBuilder::wireVarint(8) . "\x80")))
-);
-$truncatedWire = $truncatedVarintResult['canvas']['chunks'][0]['payload']['wire'] ?? array();
-$assert(false === ($truncatedWire['complete'] ?? null), 'fig-kiwi-truncated-varint-incomplete');
-$assert(true === ($truncatedWire['truncated'] ?? null), 'fig-kiwi-truncated-varint-flag');
-$assert('truncated_varint_value' === ($truncatedWire['reason'] ?? null), 'fig-kiwi-truncated-varint-reason');
-
-$unsupportedWireResult = ( new FigKiwiParser() )->parse(
-    SyntheticFigKiwiFixtureBuilder::canvas(array(SyntheticFigKiwiFixtureBuilder::zlibChunk(SyntheticFigKiwiFixtureBuilder::wireVarint(11) . 'tail')))
-);
-$unsupportedWire = $unsupportedWireResult['canvas']['chunks'][0]['payload']['wire'] ?? array();
-$assert(false === ($unsupportedWire['complete'] ?? null), 'fig-kiwi-unsupported-wire-incomplete');
-$assert(false === ($unsupportedWire['truncated'] ?? null), 'fig-kiwi-unsupported-wire-not-truncated');
-$assert('unsupported_wire_type' === ($unsupportedWire['reason'] ?? null), 'fig-kiwi-unsupported-wire-reason');
-
-$multiCandidateFixture = SyntheticFigKiwiFixtureBuilder::figArchive(
-    SyntheticFigKiwiFixtureBuilder::canvas(array(
-        SyntheticFigKiwiFixtureBuilder::jsonZlibChunk(array('metadata' => array('ignored' => true))),
-        SyntheticFigKiwiFixtureBuilder::jsonZlibChunk(SyntheticFigKiwiFixtureBuilder::nodeChangesPayload('First Candidate')),
-        SyntheticFigKiwiFixtureBuilder::jsonZlibChunk(SyntheticFigKiwiFixtureBuilder::nodeChangesPayload('Second Candidate')),
-    ))
-);
-$multiCandidateResult = blocks_engine_figma_transformer_transform_file($multiCandidateFixture);
-@unlink($multiCandidateFixture);
-$multiCandidateHtml = $fileContent($multiCandidateResult, 'index.html');
-$assert('success' === ($multiCandidateResult['status'] ?? null), 'fig-kiwi-multiple-candidates-transform-success');
-$assert(str_contains($multiCandidateHtml, 'First Candidate First'), 'fig-kiwi-multiple-candidates-renders-first-scenegraph');
-$assert(! str_contains($multiCandidateHtml, 'Second Candidate First'), 'fig-kiwi-multiple-candidates-stops-after-first-scenegraph');
-
-$nodeChangesResult = blocks_engine_figma_transformer_transform_scenegraph(array(
-    'name'         => 'Node Changes Fixture',
-    'NODE_CHANGES' => array(
-        '3:1' => array(
-            'node' => array(
-                'id'                  => '3:1',
-                'type'                => 'FRAME',
-                'name'                => 'Landing',
-                'absoluteBoundingBox' => array('x' => 0, 'y' => 0),
-                'children'            => array(
-                    array(
-                        'id'                  => '3:3',
-                        'type'                => 'TEXT',
-                        'name'                => 'Body',
-                        'characters'          => 'Second',
-                        'absoluteBoundingBox' => array('x' => 0, 'y' => 120),
-                    ),
-                    array(
-                        'id'                  => '3:2',
-                        'type'                => 'TEXT',
-                        'name'                => 'Heading',
-                        'characters'          => 'First',
-                        'absoluteBoundingBox' => array('x' => 0, 'y' => 20),
-                    ),
-                    array(
-                        'id'    => '3:4',
-                        'type'  => 'RECTANGLE',
-                        'name'  => 'Photo',
-                        'fills' => array(
-                            array('type' => 'IMAGE', 'imageRef' => 'image-hash-1'),
-                        ),
-                    ),
-                ),
-            ),
-        ),
-    ),
-));
-
-$nodeChangesHtml = (string) ($nodeChangesResult['files'][0]['content'] ?? '');
-$scenegraphReport = $nodeChangesResult['source_reports']['figma']['scenegraph'] ?? array();
-
-$assert('success' === ($nodeChangesResult['status'] ?? null), 'node-changes-transform-success');
-$assert(4 === ($nodeChangesResult['metrics']['node_count'] ?? null), 'node-changes-node-count');
-$assert(2 === ($nodeChangesResult['metrics']['text_node_count'] ?? null), 'node-changes-text-count');
-$assert(1 === ($nodeChangesResult['metrics']['asset_reference_count'] ?? null), 'node-changes-asset-count');
-$assert('3:1' === ($scenegraphReport['selected_frame_id'] ?? null), 'node-changes-selected-frame');
-$assert(false !== strpos($nodeChangesHtml, 'First') && false !== strpos($nodeChangesHtml, 'Second'), 'node-changes-html-text');
-$assert(strpos($nodeChangesHtml, 'First') < strpos($nodeChangesHtml, 'Second'), 'node-changes-stable-child-sort');
-
-$layerOrderResult = blocks_engine_figma_transformer_transform_scenegraph(array(
-    'name'         => 'Layer Order Fixture',
-    'NODE_CHANGES' => array(
-        'layer:root' => array(
-            'node' => array(
-                'id'          => 'layer:root',
-                'type'        => 'FRAME',
-                'name'        => 'Layer root',
-                'width'       => 300,
-                'height'      => 200,
-                'resizeToFit' => true,
-                'children'    => array(
-                    array(
-                        'id'          => 'layer:top',
-                        'type'        => 'RECTANGLE',
-                        'name'        => 'Top bubble',
-                        'width'       => 100,
-                        'height'      => 40,
-                        'parentIndex' => array('position' => 'b'),
-                    ),
-                    array(
-                        'id'          => 'layer:bottom',
-                        'type'        => 'RECTANGLE',
-                        'name'        => 'Bottom image',
-                        'width'       => 200,
-                        'height'      => 120,
-                        'parentIndex' => array('position' => 'a'),
-                    ),
-                ),
-            ),
-        ),
-    ),
-));
-$layerOrderHtml = $fileContent($layerOrderResult, 'index.html');
-$assert(strpos($layerOrderHtml, 'data-figma-node-id="layer:bottom"') < strpos($layerOrderHtml, 'data-figma-node-id="layer:top"'), 'freeform-layer-order-uses-parent-index-position');
-
-$overflowWrapperResult = blocks_engine_figma_transformer_transform_scenegraph(array(
-    'name'  => 'Overflow Wrapper Fixture',
-    'nodes' => array(
-        array(
-            'id'       => 'overflow:root',
-            'type'     => 'FRAME',
-            'name'     => 'Overflow Root',
-            'width'    => 100,
-            'height'   => 100,
-            'children' => array(
-                array(
-                    'id'       => 'overflow:wrapper',
-                    'type'     => 'FRAME',
-                    'name'     => 'Overflow Wrapper',
-                    'width'    => 1,
-                    'height'   => 10,
-                    'children' => array(
-                        array(
-                            'id'     => 'overflow:child',
-                            'type'   => 'RECTANGLE',
-                            'name'   => 'Overflow Child',
-                            'x'      => 4,
-                            'y'      => 2,
-                            'width'  => 20,
-                            'height' => 12,
-                        ),
-                    ),
-                ),
-            ),
-        ),
-    ),
-));
-$overflowWrapperCss = $fileContent($overflowWrapperResult, 'style.css');
-$assert(str_contains($overflowWrapperCss, '.figma-node-overflow-wrapper-overflow-wrapper{width:1px;height:10px;position:relative}'), 'overflow-wrapper-becomes-positioned-container');
-$assert(str_contains($overflowWrapperCss, '.figma-node-overflow-child-overflow-child{width:20px;height:12px;position:absolute;left:4px;top:2px}'), 'overflow-wrapper-child-keeps-local-position');
+blocks_engine_figma_transformer_run_kiwi_parser_contract($assert, $fileContent);
 
 $objectTransformResult = blocks_engine_figma_transformer_transform_scenegraph(array(
     'name'  => 'Object Transform Fixture',
@@ -6127,64 +5736,6 @@ if ( ! empty($failures) ) {
 
 fwrite(STDOUT, "Figma Transformer contract tests passed.\n");
 
-function blocks_engine_figma_transformer_create_fig_wrapper_fixture(): string
-{
-    $inner = tempnam(sys_get_temp_dir(), 'blocks-engine-inner-fig-');
-    $outer = tempnam(sys_get_temp_dir(), 'blocks-engine-wrapper-fig-');
-    if ( false === $inner || false === $outer ) {
-        throw new RuntimeException('Could not create temporary fig fixture paths.');
-    }
-
-    $canvas = 'fig-kiwi'
-        . pack('V', 106)
-        . blocks_engine_figma_transformer_kiwi_chunk(gzdeflate(json_encode(blocks_engine_figma_transformer_nodes_candidate_fixture(), JSON_THROW_ON_ERROR)))
-        . blocks_engine_figma_transformer_kiwi_chunk(gzdeflate('{"NODE_CHANGES":'))
-        . blocks_engine_figma_transformer_kiwi_chunk(gzdeflate(json_encode(blocks_engine_figma_transformer_node_changes_fixture(), JSON_THROW_ON_ERROR)))
-        . blocks_engine_figma_transformer_kiwi_chunk(gzdeflate('synthetic kiwi dictionary'))
-        . blocks_engine_figma_transformer_kiwi_chunk(blocks_engine_figma_transformer_zstd_fixture_payload());
-
-    $innerZip = new ZipArchive();
-    if ( true !== $innerZip->open($inner, ZipArchive::OVERWRITE) ) {
-        throw new RuntimeException('Could not open inner fig ZIP.');
-    }
-    $innerZip->addFromString('canvas.fig', $canvas);
-    $innerZip->addFromString('meta.json', json_encode(array('name' => 'Synthetic Fixture'), JSON_THROW_ON_ERROR));
-    $innerZip->addFromString('images/synthetic', 'asset');
-    $innerZip->close();
-
-    $outerZip = new ZipArchive();
-    if ( true !== $outerZip->open($outer, ZipArchive::OVERWRITE) ) {
-        throw new RuntimeException('Could not open wrapper fig ZIP.');
-    }
-    $outerZip->addFromString('inner.fig', (string) file_get_contents($inner));
-    $outerZip->close();
-
-    @unlink($inner);
-
-    return $outer;
-}
-
-function blocks_engine_figma_transformer_create_pending_decoder_fig_wrapper_fixture(): string
-{
-    $path = tempnam(sys_get_temp_dir(), 'blocks-engine-pending-fig-');
-    if ( false === $path ) {
-        throw new RuntimeException('Could not create temporary pending fig fixture path.');
-    }
-
-    $canvas = 'fig-kiwi'
-        . pack('V', 106)
-        . blocks_engine_figma_transformer_kiwi_chunk(gzdeflate('synthetic undecoded canvas payload'));
-
-    $zip = new ZipArchive();
-    if ( true !== $zip->open($path, ZipArchive::OVERWRITE) ) {
-        throw new RuntimeException('Could not open pending fig ZIP.');
-    }
-    $zip->addFromString('canvas.fig', $canvas);
-    $zip->close();
-
-    return $path;
-}
-
 function blocks_engine_figma_transformer_kiwi_chunk(string $payload): string
 {
     return pack('V', strlen($payload)) . $payload;
@@ -6221,610 +5772,4 @@ function blocks_engine_figma_transformer_kiwi_schema_field(string $name, int $ty
         . blocks_engine_figma_transformer_wire_varint_signed($type)
         . chr($isArray ? 1 : 0)
         . blocks_engine_figma_transformer_wire_varint($value);
-}
-
-function blocks_engine_figma_transformer_kiwi_schema_fixture(): string
-{
-    return blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_string('MessageType')
-        . chr(0)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('NODE_CHANGES', 0, false, 1)
-        . blocks_engine_figma_transformer_kiwi_string('Message')
-        . chr(2)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('type', 0, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('nodeChanges', -6, true, 4);
-}
-
-function blocks_engine_figma_transformer_kiwi_message_fixture(): string
-{
-    return blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_wire_varint(4)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_string('alpha')
-        . blocks_engine_figma_transformer_kiwi_string('beta')
-        . blocks_engine_figma_transformer_wire_varint(0);
-}
-
-function blocks_engine_figma_transformer_kiwi_frame_mask_schema_fixture(): string
-{
-    return blocks_engine_figma_transformer_wire_varint(3)
-        // def0: ENUM MessageType { NODE_CHANGES = 1 }
-        . blocks_engine_figma_transformer_kiwi_string('MessageType')
-        . chr(0)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('NODE_CHANGES', 0, false, 1)
-        // def1: MESSAGE NodeChange { type, name, frameMaskDisabled }
-        . blocks_engine_figma_transformer_kiwi_string('NodeChange')
-        . chr(2)
-        . blocks_engine_figma_transformer_wire_varint(3)
-        . blocks_engine_figma_transformer_kiwi_schema_field('type', -6, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('name', -6, false, 2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('frameMaskDisabled', -1, false, 3)
-        // def2: MESSAGE Message { type, nodeChanges[] }
-        . blocks_engine_figma_transformer_kiwi_string('Message')
-        . chr(2)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('type', 0, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('nodeChanges', 1, true, 2);
-}
-
-function blocks_engine_figma_transformer_kiwi_frame_mask_message_fixture(): string
-{
-    return blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_string('FRAME')
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_string('Masked Frame')
-        . blocks_engine_figma_transformer_wire_varint(3)
-        . chr(0)
-        . blocks_engine_figma_transformer_wire_varint(0)
-        . blocks_engine_figma_transformer_wire_varint(0);
-}
-
-function blocks_engine_figma_transformer_kiwi_derived_text_schema_fixture(): string
-{
-    return blocks_engine_figma_transformer_wire_varint(9)
-        // def0: ENUM MessageType { NODE_CHANGES = 1 }
-        . blocks_engine_figma_transformer_kiwi_string('MessageType')
-        . chr(0)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('NODE_CHANGES', 0, false, 1)
-        // def1: STRUCT Vector { x, y }
-        . blocks_engine_figma_transformer_kiwi_string('Vector')
-        . chr(1)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('x', -5, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('y', -5, false, 2)
-        // def2: STRUCT FontName { family, style, postscript }
-        . blocks_engine_figma_transformer_kiwi_string('FontName')
-        . chr(1)
-        . blocks_engine_figma_transformer_wire_varint(3)
-        . blocks_engine_figma_transformer_kiwi_schema_field('family', -6, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('style', -6, false, 2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('postscript', -6, false, 3)
-        // def3: STRUCT Baseline { position, width, lineY, lineHeight, lineAscent, firstCharacter, endCharacter }
-        . blocks_engine_figma_transformer_kiwi_string('Baseline')
-        . chr(1)
-        . blocks_engine_figma_transformer_wire_varint(7)
-        . blocks_engine_figma_transformer_kiwi_schema_field('position', 1, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('width', -5, false, 2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('lineY', -5, false, 3)
-        . blocks_engine_figma_transformer_kiwi_schema_field('lineHeight', -5, false, 4)
-        . blocks_engine_figma_transformer_kiwi_schema_field('lineAscent', -5, false, 5)
-        . blocks_engine_figma_transformer_kiwi_schema_field('firstCharacter', -4, false, 6)
-        . blocks_engine_figma_transformer_kiwi_schema_field('endCharacter', -4, false, 7)
-        // def4: STRUCT Glyph { commandsBlob, position, fontSize, firstCharacter, endCharacter, advance, rotation, styleID }
-        . blocks_engine_figma_transformer_kiwi_string('Glyph')
-        . chr(1)
-        . blocks_engine_figma_transformer_wire_varint(8)
-        . blocks_engine_figma_transformer_kiwi_schema_field('commandsBlob', -4, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('position', 1, false, 2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('fontSize', -5, false, 3)
-        . blocks_engine_figma_transformer_kiwi_schema_field('firstCharacter', -4, false, 4)
-        . blocks_engine_figma_transformer_kiwi_schema_field('endCharacter', -4, false, 5)
-        . blocks_engine_figma_transformer_kiwi_schema_field('advance', -5, false, 6)
-        . blocks_engine_figma_transformer_kiwi_schema_field('rotation', -5, false, 7)
-        . blocks_engine_figma_transformer_kiwi_schema_field('styleID', -4, false, 8)
-        // def5: STRUCT FontMetaData { key, fontLineHeight, fontWeight }
-        . blocks_engine_figma_transformer_kiwi_string('FontMetaData')
-        . chr(1)
-        . blocks_engine_figma_transformer_wire_varint(3)
-        . blocks_engine_figma_transformer_kiwi_schema_field('key', 2, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('fontLineHeight', -5, false, 2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('fontWeight', -3, false, 3)
-        // def6: STRUCT DerivedTextData { layoutSize, baselines[], glyphs[], fontMetaData }
-        . blocks_engine_figma_transformer_kiwi_string('DerivedTextData')
-        . chr(1)
-        . blocks_engine_figma_transformer_wire_varint(4)
-        . blocks_engine_figma_transformer_kiwi_schema_field('layoutSize', 1, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('baselines', 3, true, 2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('glyphs', 4, true, 3)
-        . blocks_engine_figma_transformer_kiwi_schema_field('fontMetaData', 5, false, 4)
-        // def7: MESSAGE NodeChange { type, name, derivedTextData }
-        . blocks_engine_figma_transformer_kiwi_string('NodeChange')
-        . chr(2)
-        . blocks_engine_figma_transformer_wire_varint(3)
-        . blocks_engine_figma_transformer_kiwi_schema_field('type', -6, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('name', -6, false, 2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('derivedTextData', 6, false, 3)
-        // def8: MESSAGE Message { type, nodeChanges[] }
-        . blocks_engine_figma_transformer_kiwi_string('Message')
-        . chr(2)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('type', 0, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('nodeChanges', 7, true, 2);
-}
-
-function blocks_engine_figma_transformer_kiwi_derived_text_message_fixture(): string
-{
-    return blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_string('TEXT')
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_string('Glyph Text')
-        . blocks_engine_figma_transformer_wire_varint(3)
-        // DerivedTextData.layoutSize.
-        . blocks_engine_figma_transformer_kiwi_varfloat(120.0)
-        . blocks_engine_figma_transformer_kiwi_varfloat(32.0)
-        // DerivedTextData.baselines[].
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_varfloat(0.0)
-        . blocks_engine_figma_transformer_kiwi_varfloat(22.0)
-        . blocks_engine_figma_transformer_kiwi_varfloat(120.0)
-        . blocks_engine_figma_transformer_kiwi_varfloat(22.0)
-        . blocks_engine_figma_transformer_kiwi_varfloat(24.0)
-        . blocks_engine_figma_transformer_kiwi_varfloat(18.0)
-        . blocks_engine_figma_transformer_wire_varint(0)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        // DerivedTextData.glyphs[].
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_wire_varint(7)
-        . blocks_engine_figma_transformer_kiwi_varfloat(1.0)
-        . blocks_engine_figma_transformer_kiwi_varfloat(22.0)
-        . blocks_engine_figma_transformer_kiwi_varfloat(20.0)
-        . blocks_engine_figma_transformer_wire_varint(0)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_varfloat(0.5)
-        . blocks_engine_figma_transformer_kiwi_varfloat(0.0)
-        . blocks_engine_figma_transformer_wire_varint(9)
-        // DerivedTextData.fontMetaData.
-        . blocks_engine_figma_transformer_kiwi_string('Inter')
-        . blocks_engine_figma_transformer_kiwi_string('Bold')
-        . blocks_engine_figma_transformer_kiwi_string('Inter-Bold')
-        . blocks_engine_figma_transformer_kiwi_varfloat(24.0)
-        . blocks_engine_figma_transformer_wire_varint_signed(700)
-        . blocks_engine_figma_transformer_wire_varint(0)
-        . blocks_engine_figma_transformer_wire_varint(0);
-}
-
-/**
- * Kiwi schema (version-106 shape) that defines a dev-status enum plus a
- * NodeChange/Message pair carrying `sectionStatus`. Proves the field-policy
- * extension (#280) reads the status through the REAL generic decoder.
- */
-function blocks_engine_figma_transformer_kiwi_dev_status_schema_fixture(): string
-{
-    return blocks_engine_figma_transformer_wire_varint(3)
-        // def0: ENUM SectionStatus { BUILD = 1, COMPLETED = 2 } (real Figma enum; BUILD = "Ready for dev")
-        . blocks_engine_figma_transformer_kiwi_string('SectionStatus')
-        . chr(0)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('BUILD', 0, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('COMPLETED', 0, false, 2)
-        // def1: MESSAGE NodeChange { type, name, sectionStatus }
-        . blocks_engine_figma_transformer_kiwi_string('NodeChange')
-        . chr(2)
-        . blocks_engine_figma_transformer_wire_varint(3)
-        . blocks_engine_figma_transformer_kiwi_schema_field('type', -6, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('name', -6, false, 2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('sectionStatus', 0, false, 3)
-        // def2: MESSAGE Message { type, nodeChanges[] }
-        . blocks_engine_figma_transformer_kiwi_string('Message')
-        . chr(2)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('type', -6, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('nodeChanges', 1, true, 2);
-}
-
-/**
- * Kiwi message for {@see blocks_engine_figma_transformer_kiwi_dev_status_schema_fixture()}:
- * one NodeChange of type SECTION with sectionStatus = COMPLETED (enum value 2).
- */
-function blocks_engine_figma_transformer_kiwi_dev_status_message_fixture(): string
-{
-    return blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_string('DOCUMENT')
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_string('SECTION')
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_string('Completed Section')
-        . blocks_engine_figma_transformer_wire_varint(3)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_wire_varint(0)
-        . blocks_engine_figma_transformer_wire_varint(0);
-}
-
-/**
- * Encode a float using Kiwi's varFloat format (the inverse of
- * {@see FigKiwiByteReader::readVarFloat()}): rotate the IEEE-754 bits left by 9
- * so the sign/exponent land in the low byte, and collapse exact 0.0 to one byte.
- */
-function blocks_engine_figma_transformer_kiwi_varfloat(float $value): string
-{
-    $bits = unpack('V', pack('f', $value));
-    $raw = is_array($bits) ? (int) $bits[1] : 0;
-    if ( 0 === $raw ) {
-        return chr(0);
-    }
-
-    $rotated = ( ( $raw << 9 ) & 0xffffffff ) | ( ( $raw >> 23 ) & 0x1ff );
-    return pack('V', $rotated);
-}
-
-/**
- * Encode a float in the Kiwi "compressed float" form that
- * {@see FigKiwiByteReader::readVarFloat()} consumes: a single 0 byte for zero,
- * otherwise the IEEE-754 bits rotated right by 23 and emitted little-endian.
- */
-function blocks_engine_figma_transformer_kiwi_float(float $value): string
-{
-    if ( 0.0 === $value ) {
-        return chr(0);
-    }
-
-    $bits = unpack('V', pack('f', $value));
-    $ieee = is_array($bits) ? (int) $bits[1] : 0;
-    // Inverse of the decoder's rotate-left-by-23 so the round trip reproduces $value.
-    $rotated = ( ( $ieee << 9 ) & 0xffffffff ) | ( ( $ieee >> 23 ) & 0x1ff );
-    return pack('V', $rotated);
-}
-
-/**
- * Kiwi schema defining a StrokeAlign enum plus a NodeChange/Message pair that
- * carries `strokeWeight` (float), `strokeAlign` (enum) and `dashPattern`
- * (float[]). Proves the field-policy extension (#328) reads stroke geometry
- * through the REAL generic decoder instead of dropping it at the whitelist.
- */
-function blocks_engine_figma_transformer_kiwi_stroke_schema_fixture(): string
-{
-    return blocks_engine_figma_transformer_wire_varint(3)
-        // def0: ENUM StrokeAlign { CENTER = 1, INSIDE = 2, OUTSIDE = 3 }
-        . blocks_engine_figma_transformer_kiwi_string('StrokeAlign')
-        . chr(0)
-        . blocks_engine_figma_transformer_wire_varint(3)
-        . blocks_engine_figma_transformer_kiwi_schema_field('CENTER', 0, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('INSIDE', 0, false, 2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('OUTSIDE', 0, false, 3)
-        // def1: MESSAGE NodeChange { type, name, strokeWeight, strokeAlign, dashPattern[] }
-        . blocks_engine_figma_transformer_kiwi_string('NodeChange')
-        . chr(2)
-        . blocks_engine_figma_transformer_wire_varint(5)
-        . blocks_engine_figma_transformer_kiwi_schema_field('type', -6, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('name', -6, false, 2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('strokeWeight', -5, false, 3)
-        . blocks_engine_figma_transformer_kiwi_schema_field('strokeAlign', 0, false, 4)
-        . blocks_engine_figma_transformer_kiwi_schema_field('dashPattern', -5, true, 5)
-        // def2: MESSAGE Message { type, nodeChanges[] }
-        . blocks_engine_figma_transformer_kiwi_string('Message')
-        . chr(2)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_schema_field('type', -6, false, 1)
-        . blocks_engine_figma_transformer_kiwi_schema_field('nodeChanges', 1, true, 2);
-}
-
-/**
- * Kiwi message for {@see blocks_engine_figma_transformer_kiwi_stroke_schema_fixture()}:
- * one RECTANGLE NodeChange with strokeWeight = 3, strokeAlign = INSIDE and a
- * dashPattern of [4, 2].
- */
-function blocks_engine_figma_transformer_kiwi_stroke_message_fixture(): string
-{
-    return blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_string('DOCUMENT')
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_wire_varint(1)
-        // NodeChange[0]
-        . blocks_engine_figma_transformer_wire_varint(1)
-        . blocks_engine_figma_transformer_kiwi_string('RECTANGLE')
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_string('Bordered Rect')
-        . blocks_engine_figma_transformer_wire_varint(3)
-        . blocks_engine_figma_transformer_kiwi_float(3.0)
-        . blocks_engine_figma_transformer_wire_varint(4)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_wire_varint(5)
-        . blocks_engine_figma_transformer_wire_varint(2)
-        . blocks_engine_figma_transformer_kiwi_float(4.0)
-        . blocks_engine_figma_transformer_kiwi_float(2.0)
-        . blocks_engine_figma_transformer_wire_varint(0)
-        . blocks_engine_figma_transformer_wire_varint(0);
-}
-
-/**
- * Kiwi schema for component-property text overrides. The field names and nested
- * shape match the FSE Pilot production schema: instance assignments carry
- * ComponentPropAssignment.defID/value.textValue.characters while master text
- * nodes carry ComponentPropRef.defID/componentPropNodeField = TEXT_DATA.
- */
-function blocks_engine_figma_transformer_kiwi_component_prop_schema_fixture(): string
-{
-    $field = 'blocks_engine_figma_transformer_kiwi_schema_field';
-    $str = 'blocks_engine_figma_transformer_kiwi_string';
-    $varint = 'blocks_engine_figma_transformer_wire_varint';
-
-    return $varint(10)
-        // def0: ENUM ComponentPropNodeField { TEXT_DATA = 1 }
-        . $str('ComponentPropNodeField') . chr(0) . $varint(1)
-        . $field('TEXT_DATA', 0, false, 1)
-        // def1: STRUCT GUID { sessionID, localID }
-        . $str('GUID') . chr(1) . $varint(2)
-        . $field('sessionID', -4, false, 1)
-        . $field('localID', -4, false, 2)
-        // def2: MESSAGE TextData { characters }
-        . $str('TextData') . chr(2) . $varint(1)
-        . $field('characters', -6, false, 1)
-        // def3: MESSAGE ComponentPropValue { textValue }
-        . $str('ComponentPropValue') . chr(2) . $varint(1)
-        . $field('textValue', 2, false, 2)
-        // def4: MESSAGE VariableAnyValue { textDataValue }
-        . $str('VariableAnyValue') . chr(2) . $varint(1)
-        . $field('textDataValue', 2, false, 10)
-        // def5: MESSAGE VariableData { value, dataType, resolvedDataType }
-        . $str('VariableData') . chr(2) . $varint(3)
-        . $field('value', 4, false, 1)
-        . $field('dataType', -6, false, 2)
-        . $field('resolvedDataType', -6, false, 3)
-        // def6: MESSAGE ComponentPropAssignment { defID, value, varValue }
-        . $str('ComponentPropAssignment') . chr(2) . $varint(3)
-        . $field('defID', 1, false, 1)
-        . $field('value', 3, false, 2)
-        . $field('varValue', 5, false, 3)
-        // def7: MESSAGE ComponentPropRef { defID, componentPropNodeField }
-        . $str('ComponentPropRef') . chr(2) . $varint(2)
-        . $field('defID', 1, false, 2)
-        . $field('componentPropNodeField', 0, false, 4)
-        // def8: MESSAGE NodeChange { type, name, componentPropAssignments[], componentPropRefs[], pluginData }
-        . $str('NodeChange') . chr(2) . $varint(5)
-        . $field('type', -6, false, 1)
-        . $field('name', -6, false, 2)
-        . $field('componentPropAssignments', 6, true, 3)
-        . $field('componentPropRefs', 7, true, 4)
-        . $field('pluginData', -2, true, 5)
-        // def9: MESSAGE Message { type, nodeChanges[] }
-        . $str('Message') . chr(2) . $varint(2)
-        . $field('type', -6, false, 1)
-        . $field('nodeChanges', 8, true, 2);
-}
-
-/**
- * Kiwi message for {@see blocks_engine_figma_transformer_kiwi_component_prop_schema_fixture()}.
- */
-function blocks_engine_figma_transformer_kiwi_component_prop_message_fixture(): string
-{
-    $str = 'blocks_engine_figma_transformer_kiwi_string';
-    $v = 'blocks_engine_figma_transformer_wire_varint';
-
-    $defId = $v(9) . $v(10); // GUID struct body; structs carry no terminator.
-    $textData = $v(1) . $str('Selected label') . $v(0);
-    $value = $v(2) . $textData . $v(0);
-    $varTextData = $v(1) . $str('Selected label via variable') . $v(0);
-    $varAnyValue = $v(10) . $varTextData . $v(0);
-    $varValue = $v(1) . $varAnyValue
-        . $v(2) . $str('TEXT')
-        . $v(3) . $str('TEXT')
-        . $v(0);
-    $assignment = $v(1) . $defId
-        . $v(2) . $value
-        . $v(3) . $varValue
-        . $v(0);
-    $ref = $v(2) . $defId
-        . $v(4) . $v(1)
-        . $v(0);
-    $node = $v(1) . $str('INSTANCE')
-        . $v(2) . $str('Component Property Instance')
-        . $v(3) . $v(1) . $assignment
-        . $v(4) . $v(1) . $ref
-        . $v(5) . $v(3) . 'raw'
-        . $v(0);
-
-    return $v(1) . $str('DOCUMENT')
-        . $v(2) . $v(1) . $node
-        . $v(0);
-}
-
-/**
- * Kiwi schema (version-106 shape) that defines the prototype-link surface (#328):
- * a `Hyperlink` struct plus the `PrototypeInteraction`/`PrototypeAction` graph
- * hanging off `NodeChange`, with the real Figma `ConnectionType`/`NavigationType`
- * enums. Proves the field-policy additions read links through the REAL decoder.
- */
-function blocks_engine_figma_transformer_kiwi_link_schema_fixture(): string
-{
-    $field = 'blocks_engine_figma_transformer_kiwi_schema_field';
-    $str = 'blocks_engine_figma_transformer_kiwi_string';
-    $varint = 'blocks_engine_figma_transformer_wire_varint';
-
-    return $varint(10)
-        // def0: STRUCT GUID { sessionID, localID }
-        . $str('GUID') . chr(1) . $varint(2)
-        . $field('sessionID', -4, false, 1)
-        . $field('localID', -4, false, 2)
-        // def1: ENUM InteractionType { ON_CLICK = 0 }
-        . $str('InteractionType') . chr(0) . $varint(1)
-        . $field('ON_CLICK', 0, false, 0)
-        // def2: ENUM ConnectionType { NONE = 0, INTERNAL_NODE = 1, URL = 2 }
-        . $str('ConnectionType') . chr(0) . $varint(3)
-        . $field('NONE', 0, false, 0)
-        . $field('INTERNAL_NODE', 0, false, 1)
-        . $field('URL', 0, false, 2)
-        // def3: ENUM NavigationType { NAVIGATE = 0 }
-        . $str('NavigationType') . chr(0) . $varint(1)
-        . $field('NAVIGATE', 0, false, 0)
-        // def4: MESSAGE PrototypeEvent { interactionType }
-        . $str('PrototypeEvent') . chr(2) . $varint(1)
-        . $field('interactionType', 1, false, 1)
-        // def5: MESSAGE PrototypeAction { transitionNodeID, connectionType, connectionURL, navigationType }
-        . $str('PrototypeAction') . chr(2) . $varint(4)
-        . $field('transitionNodeID', 0, false, 1)
-        . $field('connectionType', 2, false, 7)
-        . $field('connectionURL', -6, false, 8)
-        . $field('navigationType', 3, false, 10)
-        // def6: MESSAGE PrototypeInteraction { event, actions[] }
-        . $str('PrototypeInteraction') . chr(2) . $varint(2)
-        . $field('event', 4, false, 2)
-        . $field('actions', 5, true, 3)
-        // def7: MESSAGE Hyperlink { url, guid }
-        . $str('Hyperlink') . chr(2) . $varint(2)
-        . $field('url', -6, false, 1)
-        . $field('guid', 0, false, 2)
-        // def8: MESSAGE NodeChange { type, name, hyperlink, prototypeInteractions[] }
-        . $str('NodeChange') . chr(2) . $varint(4)
-        . $field('type', -6, false, 1)
-        . $field('name', -6, false, 2)
-        . $field('hyperlink', 7, false, 4)
-        . $field('prototypeInteractions', 6, true, 5)
-        // def9: MESSAGE Message { type, nodeChanges[] }
-        . $str('Message') . chr(2) . $varint(2)
-        . $field('type', -6, false, 1)
-        . $field('nodeChanges', 8, true, 2);
-}
-
-/**
- * Kiwi message for {@see blocks_engine_figma_transformer_kiwi_link_schema_fixture()}:
- * a TEXT NodeChange with a URL `hyperlink`, and a FRAME NodeChange with an
- * ON_CLICK `prototypeInteractions` entry whose actions carry a URL connection
- * and a node-navigation connection (GUID destination).
- */
-function blocks_engine_figma_transformer_kiwi_link_message_fixture(): string
-{
-    $str = 'blocks_engine_figma_transformer_kiwi_string';
-    $v = 'blocks_engine_figma_transformer_wire_varint';
-
-    // Hyperlink { url = "https://example.com/about" }
-    $hyperlink = $v(1) . $str('https://example.com/about') . $v(0);
-
-    // PrototypeEvent { interactionType = ON_CLICK (0) }
-    $event = $v(1) . $v(0) . $v(0);
-
-    // PrototypeAction { connectionType = URL (2), connectionURL }
-    $actionUrl = $v(7) . $v(2)
-        . $v(8) . $str('https://example.com/cta')
-        . $v(0);
-
-    // PrototypeAction { transitionNodeID = GUID{7,42}, connectionType = INTERNAL_NODE (1), navigationType = NAVIGATE (0) }
-    $actionNode = $v(1) . $v(7) . $v(42) // GUID struct body (sessionID, localID; structs carry no terminator)
-        . $v(7) . $v(1)
-        . $v(10) . $v(0)
-        . $v(0);
-
-    // PrototypeInteraction { event, actions = [actionUrl, actionNode] }
-    $interaction = $v(2) . $event
-        . $v(3) . $v(2) . $actionUrl . $actionNode
-        . $v(0);
-
-    // NodeChange A: TEXT { hyperlink }
-    $nodeA = $v(1) . $str('TEXT')
-        . $v(2) . $str('External link')
-        . $v(4) . $hyperlink
-        . $v(0);
-
-    // NodeChange B: FRAME { prototypeInteractions = [interaction] }
-    $nodeB = $v(1) . $str('FRAME')
-        . $v(2) . $str('CTA')
-        . $v(5) . $v(1) . $interaction
-        . $v(0);
-
-    // Message { type = "DOCUMENT", nodeChanges = [nodeA, nodeB] }
-    return $v(1) . $str('DOCUMENT')
-        . $v(2) . $v(2) . $nodeA . $nodeB
-        . $v(0);
-}
-
-/**
- * @return array<string, mixed>
- */
-function blocks_engine_figma_transformer_node_changes_fixture(): array
-{
-    return array(
-        'name'         => 'Decoded Node Changes Fixture',
-        'NODE_CHANGES' => array(
-            '4:1' => array(
-                'node' => array(
-                    'id'       => '4:1',
-                    'type'     => 'FRAME',
-                    'name'     => 'Decoded Landing',
-                    'children' => array(
-                        array(
-                            'id'         => '4:2',
-                            'type'       => 'TEXT',
-                            'name'       => 'Heading',
-                            'characters' => 'Decoded First',
-                        ),
-                        array(
-                            'id'         => '4:3',
-                            'type'       => 'TEXT',
-                            'name'       => 'Body',
-                            'characters' => 'Decoded Second',
-                        ),
-                        array(
-                            'id'   => '4:4',
-                            'type' => 'RECTANGLE',
-                            'name' => 'Decoded Photo',
-                            'fills' => array(
-                                array('type' => 'IMAGE', 'imageHash' => 'synthetic'),
-                            ),
-                        ),
-                    ),
-                ),
-            ),
-        ),
-    );
-}
-
-/**
- * @return array<string, mixed>
- */
-function blocks_engine_figma_transformer_nodes_candidate_fixture(): array
-{
-    return array(
-        'name'  => 'Lower Priority Nodes Candidate',
-        'nodes' => array(
-            array(
-                'id'       => 'candidate:1',
-                'type'     => 'FRAME',
-                'name'     => 'Lower Priority Frame',
-                'children' => array(
-                    array(
-                        'id'         => 'candidate:2',
-                        'type'       => 'TEXT',
-                        'name'       => 'Lower Priority Text',
-                        'characters' => 'This earlier payload must not be selected.',
-                    ),
-                ),
-            ),
-        ),
-    );
-}
-
-function blocks_engine_figma_transformer_zstd_fixture_payload(): string
-{
-    if ( function_exists('zstd_compress') ) {
-        $compressed = zstd_compress('synthetic zstd payload');
-        if ( false !== $compressed ) {
-            return $compressed;
-        }
-    }
-
-    return "\x28\xb5\x2f\xfd" . 'synthetic-zstd-frame';
 }
